Added a way to make it easier to send TODO reminders elsewhere than slack:
- End goal is to make SmartTodo pluggable for a user and allow him
  to implement his own dispatcher (Gmail, Microsoft workspace ...) to
  send todo reminders to.
