+++ test/smart_todo/cli_test.rb
@@ -5,41 +5,6 @@ require 'tempfile'
 
 module SmartTodo
   class CLITest < Minitest::Test
-    def test_when_all_mandatory_options_are_passed
-      cli = CLI.new
-      paths = cli.run([__FILE__, '--slack_token', '123', '--fallback_channel', '#general"'])
-
-      assert_equal([__FILE__], paths)
-    end
-
-    def test_when_slack_token_option_is_missing
-      cli = CLI.new
-
-      error = assert_raises(ArgumentError) do
-        cli.run([__FILE__, '--fallback_channel', '"#general"'])
-      end
-      assert_equal('Missing :slack_token', error.message)
-    end
-
-    def test_when_slack_token_option_is_in_the_environment
-      ENV['SMART_TODO_SLACK_TOKEN'] = '123'
-      cli = CLI.new
-      paths = cli.run([__FILE__, '--fallback_channel', '"#general"'])
-
-      assert_equal([__FILE__], paths)
-    ensure
-      ENV.delete('SMART_TODO_SLACK_TOKEN')
-    end
-
-    def test_when_fallback_channel_is_missing
-      cli = CLI.new
-
-      error = assert_raises(ArgumentError) do
-        cli.run([__FILE__, '--slack_token', '123'])
-      end
-      assert_equal('Missing :fallback_channel', error.message)
-    end
-
     def test_adds_current_directory_if_none_is_passed
       cli = CLI.new
 
@@ -64,7 +29,7 @@ module SmartTodo
       mock.expect(:dispatch, nil)
 
       generate_ruby_file(ruby_code) do |file|
-        Dispatcher.stub(:new, mock) do
+        Dispatchers::Slack.stub(:new, mock) do
           cli.run([file.path, '--slack_token', '123', '--fallback_channel', '#general"'])
         end
       end
